feat: Add RenderManager
